@@ -21,36 +21,28 @@ use Doctrine\ORM\Mapping as ORM;
 class Participant
 {
     /**
-     * @var int
-     *
      * @ORM\Column(type="integer")
      *
      * @ORM\Id
      *
      * @ORM\GeneratedValue(strategy="IDENTITY")
      */
-    private $id;
+    private int $id;
 
     /**
-     * @var Round
-     *
      * @ORM\ManyToOne(targetEntity="Round", inversedBy="participants")
      */
-    private $round;
+    private Round $round;
 
     /**
-     * @var Player
-     *
      * @ORM\ManyToOne(targetEntity="Player")
      */
-    private $player;
+    private Player $player;
 
     /**
-     * @var int
-     *
      * @ORM\Column(type="integer")
      */
-    private $points;
+    private int $points;
 
     public function __construct(Round $round, Player $player, int $points)
     {
